src/requirements.txt
asgiref==3.7.2
backports.entry-points-selectable==1.1.1
barcode==1.0.2
certifi==2022.5.18
charset-normalizer==2.0.12
chartjs==1.2
concurrency==0.1.0
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.4
Django==2.2.14
django-appconf==1.0.5
django-betterforms==1.2
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==0.0.7
django-concurrency==2.4
django-crispy-forms==1.13.0
django-imagekit==4.1.0
django-import-export==2.5.0
django-mysql==3.8.1
django-session-timeout==0.1.0
django-webpack-loader==0.7.0
django-widget-tweaks==1.4.3
django-widgets-improved==1.5.0
django-wkhtmltopdf==3.2.0
djangorestframework==3.12.4
et-xmlfile==1.0.1
filelock==3.4.0
futures==3.0.5
html5lib==1.1
idna==3.3
imgkit==1.2.3
importlib-metadata==4.9.0
jdcal==1.4.1
Levenshtein==0.20.9
MarkupPy==1.14
mysqlclient==2.0.1
odfpy==1.4.1
openpyxl==3.0.6
payjp==0.1.0
pdfkit==0.6.1
pilkit==2.0
Pillow==9.1.1
platformdirs==2.4.0
PyPDF2==1.26.0
python-barcode==0.14.0
python-dateutil==2.8.1
python-Levenshtein==0.20.9
python-xz==0.5.0
pytz==2020.1
PyYAML==5.4.1
rapidfuzz==2.13.7
reportlab==3.6.9
requests==2.27.1
six==1.15.0
sqlparse==0.3.1
stripe==5.2.0
tablib==3.0.0
tk==0.1.0
typing_extensions==4.7.1
tzdata==2023.3
ua-parser==0.10.0
urllib3==1.26.9
user-agents==2.1
virtualenv==20.10.0
webencodings==0.5.1
xhtml2pdf==0.2.3
xlrd==2.0.1
xlwt==1.3.0
zipp==3.6.0